net: context: Fix the ICMP error on udp
This commit fixes the #xxxx.

When receiving a UDP packet, net_conn_input() searches for a
matching connection within `conn_used`.

However, when receiving UDP packets simultaneously from multiple
clients, we may encounter a situation where the connection that was
supposed to be bound cannot be found within `conn_used`, and raise
the ICMP error.

This is because, within recv_udp(), to avoid the failure of
bind_default(), we temporarily remove it from `conn_used` using
net_conn_unregister().

If net context already have connection handler, then it effectively
means that it's already bound to an interface/port and register a
connection. In that case, it's just a matter of updating the
callback registered in the net_context and the user data passed to
that callback. The callback function passed to
net_conn_change_callback() specifies same function passed in
net_conn_register().

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>